requirements.txt
apify-client==1.4.1
google-generativeai==0.3.1
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.1.0
tqdm==4.66.1